packages/cad/pyproject.toml
[project]
name = "openarm-cad"
version = "0.1.0"
description = "CAD asset processing and CDN deployment for OpenArm"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pybullet>=3.2.7",
    "mujoco>=3.3.5",
    "onshape-to-robot>=1.7.7",
    "boto3>=1.35.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["processors"]
